Changed the ElementType to be float instead of double of als algorithms.

+++ flink-jobs/src/main/scala/com/github/projectflink/als/ALSFlink.scala
@@ -5,6 +5,6 @@ import org.apache.flink.api.scala.{DataSet}
 
 trait ALSFlink extends ALS{
   type DS[T] = DataSet[T]
-  type ElementType = Double
+  type ElementType = Float
   type IDType = Int
 }
